requirements.txt
astroid==1.6.3
autopep8==1.3.5
Django==1.11.13
django-debug-panel==0.8.3
django-debug-toolbar==1.9.1
isort==4.3.4
lazy-object-proxy==1.3.1
lorem==0.1.1
mccabe==0.6.1
pycodestyle==2.4.0
pylint==1.8.4
pytz==2018.4
six==1.11.0
sqlparse==0.2.4
wrapt==1.10.11
